api/requirements.txt
click==7.1.2
Flas==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.1
scipy==1.5.1
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
Werkzeug==1.0.1